// src/app/master/store/models/part-mapping.model.ts
export interface StorePartMapping {
    StoreZoneDetailId: number;
    StoreBinDetailId: number;
    StorePartMappingId?: number;
    PartTypeId: number;
    PartId: number;
    StoreId: number;
    PartCode:string;
    ZoneCode:string;
    BinCode:string;
    createdBy: string;
    created?: Date;
    modifiedBy: string;
    modified?: Date;
  }
  